ft-user-profile-173317076(user profile): - restructure files
